+++ modules/infra-mq/src/main/java/com/ott/infra/mq/TranscodeConstants.java
@@ -1,0 +1,10 @@
+package com.ott.infra.mq;
+
+public final class TranscodeConstants {
+
+    public static final String EXCHANGE_NAME = "transcode.exchange";
+    public static final String ROUTING_KEY = "transcode.request";
+
+    private TranscodeConstants() {
+    }
+}
