+++ modules/infra-mq/src/main/java/com/ott/infra/mq/config/MqMessageConfig.java
@@ -1,0 +1,27 @@
+package com.ott.infra.mq.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ott.infra.mq.TranscodeMessage;
+import org.springframework.amqp.support.converter.DefaultClassMapper;
+import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
+import org.springframework.amqp.support.converter.MessageConverter;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class MqMessageConfig {
+
+    @Bean
+    public DefaultClassMapper classMapper() {
+        DefaultClassMapper classMapper = new DefaultClassMapper();
+        classMapper.setDefaultType(TranscodeMessage.class);
+        return classMapper;
+    }
+
+    @Bean
+    public MessageConverter jacksonMessageConverter(ObjectMapper objectMapper, DefaultClassMapper classMapper) {
+        Jackson2JsonMessageConverter converter = new Jackson2JsonMessageConverter(objectMapper);
+        converter.setClassMapper(classMapper);
+        return converter;
+    }
+}
