+++ settings.gradle
@@ -9,6 +9,7 @@ include(
         // Modules
         ":modules:domain",
         ":modules:infra-db",
+        ":modules:infra-mq",
         ":modules:infra-s3",
         ":modules:common-web",
         ":modules:common-security"
